Make changes
Move ModifyPlayerZoneStatistic to head gm mod.rs
Move UpdatePlayerStatistic back to client received
