+++ src/world/gm/client/tests/ModifyPlayerZoneStatistic.rs
@@ -1,0 +1,8 @@
+GameMessage::ModifyPlayerZoneStatistic(
+	ModifyPlayerZoneStatistic {
+		set: false,
+		stat_name: lu!("AchievementsCompleted"),
+		stat_value: 1,
+		zone_id: 1101,
+	},
+)
