+++ src/world/gm/client/tests/UpdatePlayerStatistic.rs
@@ -1,6 +1,6 @@
 GameMessage::UpdatePlayerStatistic(
 	UpdatePlayerStatistic {
-		update_id: StatisticID::MetersTraveled,
+		update_id: StatisticId::MetersTraveled,
 		update_value: 29,
 	},
 )
